Move LogsRequestFactory into Logs module, update coverage for nightly tests
