+++ instrumentation-tools/src/constants.py
@@ -9,7 +9,8 @@ API_SURFACE_PATHS = [
     "library/dd-sdk-android-session-replay/apiSurface",
     "library/dd-sdk-android-logs/apiSurface",
     "library/dd-sdk-android-ndk/apiSurface",
-    "library/dd-sdk-android-trace/apiSurface"
+    "library/dd-sdk-android-trace/apiSurface",
+    "library/dd-sdk-android-webview/apiSurface"
 ]
 NIGHTLY_TESTS_DIRECTORY_PATH = "instrumented/nightly-tests/src/androidTest/kotlin"
 NIGHTLY_TESTS_PACKAGE = "com/datadog/android/nightly"
